+++ src/block.ts
@@ -187,10 +187,8 @@ export abstract class Block
      */
     public getTypeFromValue(value:string):string
     {
-        let result:Array<string>;
-
         // Check for bool
-        if (value.match(/^\s*(false|true)\s*$/i) !== null) {
+        if (value.match(/^\s*(false|true)\s*$/i) !== null || value.match(/^\s*\!/i) !== null) {
             return TypeUtil.instance.getFormattedTypeByName('bool');
         }
 
@@ -205,7 +203,7 @@ export abstract class Block
         }
 
         // Check for string
-        if (value.match(/^\s*(["'])/) !== null) {
+        if (value.match(/^\s*(["'])/) !== null || value.match(/^\s*<<</) !== null) {
             return 'string';
         }
 
@@ -214,7 +212,28 @@ export abstract class Block
             return 'array';
         }
 
-        return '[type]';
+        // Check for class
+        var match = value.match(/^\s*new\s+([a-z0-9_\\\|]+)/i);
+        if (match) {
+            if (match[1] === 'class') {
+                return 'object';
+            }
+            return match[1];
+        }
+
+        // Check for closure
+        var match = value.match(/^\s*function\s*\(/i);
+        if (match) {
+            return '\\Closure';
+        }
+
+        // Check for type casting
+        var match = value.match(/^\s*\(\s*(int|integer|bool|boolean|float|double|real|string|array|object|unset)\s*\)/i);
+        if (match) {
+            return TypeUtil.instance.getFormattedTypeByName(match[1]);
+        }
+
+        return TypeUtil.instance.getUnknownType();
     }
 
     /**