+++ src/doc.ts
@@ -38,6 +38,13 @@ export class Doc
     public message:string;
 
     /**
+     * Doc inline
+     *
+     * @type {boolean}
+     */
+    public inline:boolean = false;
+
+    /**
      * Define the template for the documentor
      *
      * @type {Object}
@@ -69,6 +76,9 @@ export class Doc
         }
         if (input.message !== undefined) {
             this.message = input.message;
+        }
+        if (input.inline !== undefined) {
+            this.inline = input.inline;
         }
         if (input.params !== undefined && Array.isArray(input.params)) {
             input.params.forEach(param => {
@@ -114,7 +124,14 @@ export class Doc
         }
 
         if (this.var) {
-            varString = "@var \${###:" +this.var + "}";
+            // @var type $name
+            varString = "@var";
+            let vars = this.var.split(' ');
+            for (let index = 0; index < vars.length; index++) {
+                if (vars[index] !== '') {
+                    varString += " \${###:" + vars[index].replace('$', '\\$') + "}";
+                }
+            }
         }
 
         if (this.return && (this.return != 'void' || Config.instance.get('returnVoid'))) {
@@ -167,8 +184,14 @@ export class Doc
             templateArray.pop();
         }
 
-        let templateString:string = templateArray.join("\n");
-        templateString = "/**\n" + templateString + "\n */";
+        let templateString:string;
+
+        if (this.inline && templateArray.length === 3) {
+            templateString = "/** " + templateArray[2] + " " + templateArray[0] + " */";
+        } else {
+            templateString = templateArray.join("\n");
+            templateString = "/**\n" + templateString + "\n */";
+        }
 
         let stop = 0;
         templateString = templateString.replace(/###/gm, function():string {