+++ test/functions.test.ts
@@ -23,6 +23,9 @@ suite("Function tests", () => {
     });
 
     map.forEach(testData => {
+        if (testData.name === undefined) {
+            testData.name = testData.key;
+        }
         test("Match Test: "+ testData.name, () => {
             let func = new Function(testPositions[testData.key], editor);
             assert.equal(func.test(), true, test.name);